refactor: 헤더에서 마이페이지 이동 구현 및 스타일 변경
