+++ src/pages/MyPage/index.tsx
@@ -1,17 +1,19 @@
 import HeaderBar from '@/components/HeaderBar';
-import { Wrapper } from '@/styles/DefaultSide/styles.tsx';
+import { Wrapper } from '@/styles/DetailSide/styles.tsx';
 import SideBar from '@/components/SideBar';
 import MyInfoCard from '@/components/MyInfoCard';
 import { CenterDiv, ChangeInfoDiv, MyInfoCardDiv, MyPageDiv, MyPageTextDiv } from '@/pages/MyPage/styles.tsx';
 import { Divider } from 'antd';
 import ChangePasswordModal from '@/components/MyInfoCard/ChangePasswordModal';
 import WithdrawAccountModal from '@/components/MyInfoCard/WithdrawAccountModal';
+import SideDetailBar from '@components/SideDetailBar';
 
 const MyPage = () => {
   return (
     <>
       <HeaderBar />
       <SideBar />
+      <SideDetailBar />
       <Wrapper>
         <CenterDiv>
           <MyPageDiv>
